requirements.txt
astroid==3.3.11
black==25.1.0
click
dill==0.4.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
numpy
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
py4j==0.10.9.9
pylint==3.3.8
pyspark==4.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tomlkit==0.13.3
tzdata==2025.2
